refactor(gateway): extract entry points from run.py (slice 31 of #54962)
